backend/requirements.txt
aiofiles==23.2.1
aiogram==3.7.0
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crontab==1.0.1
et-xmlfile==1.1.0
exceptiongroup==1.2.1
freezegun==1.5.1
frozenlist==1.4.1
h11==0.14.0
idna==3.7
kombu==5.3.7
magic-filter==1.0.12
multidict==6.0.5
nest-asyncio==1.6.0
openpyxl==3.1.4
outcome==1.3.0.post0
packaging==24.1
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pydantic==2.7.4
pydantic_core==2.18.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.5
requests==2.32.3
rq==1.16.2
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.1
wsproto==1.2.0
yarl==1.9.4
